+++ libraries/Stepper/src/Stepper.h
@@ -91,7 +91,7 @@ class Stepper {
                                  int motor_pin_5);
 
     // speed setter method:
-    void setSpeed(long whatSpeed);
+    void setSpeed(float whatSpeed);
 
     // mover method:
     void step(int number_of_steps);
@@ -102,7 +102,7 @@ class Stepper {
     void stepMotor(int this_step);
 
     int direction;            // Direction of rotation
-    unsigned long step_delay; // delay between steps, in ms, based on speed
+    unsigned long step_delay; // delay between steps, in us, based on speed
     int number_of_steps;      // total number of steps this motor can take
     int pin_count;            // how many pins are in use.
     int step_number;          // which step the motor is on
